requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Babel==2.9.1
backports.zoneinfo==0.2.1
Django==4.0.4
django-cors-headers==3.8.0
django-filter==2.4.0
django-money==2.1.1
djangorestframework==3.12.4
flake8==3.9.2
gunicorn==20.1.0
Markdown==3.3.4
mccabe==0.6.1
Pillow==8.3.2
psycopg2-binary==2.9.1
py-moneyed==1.2
pyclean==2.2.0
pycodestyle==2.7.0
pyflakes==2.3.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
